+++ web/api/views.py
@@ -1403,7 +1403,7 @@ def tasks_iocs(request, task_id, detail=None):
     del data["info"]["custom"]
     # The machines key won't exist in cases where an x64 binary is submitted
     # when there are no x64 machines.
-    if "machine" in data["info"] and data["info"]["machine"]:
+    if data.get("info", {}).get("machine", {}) and isinstance(data["info"]["machine"], dict):
         del data["info"]["machine"]["manager"]
         del data["info"]["machine"]["label"]
         del data["info"]["machine"]["id"]